third_party/blink/web_tests/fast/events/onunload-not-on-body-expected.txt

CONSOLE WARNING: line 1: Triggering window.alert from cross origin iframes has been deprecated and will be removed in the future.
CONSOLE ERROR: Blocked alert('unload') during unload.
You should only see one unload alert appear.
